backend/pyproject.toml
[project]
name = "voxelink-backend"
version = "0.1.0"
description = ""
authors = []
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9,<4"
dependencies = [
    "fastapi (~=0.115)",
    "loguru (~=0.7)",
    "mem0ai (~=0.1.114)",
    "openai (~=1.76)",
    "python-dotenv (~=1.0)",
    "python-multipart (~=0.0)",
    "tiktoken (~=0.9)",
    "tortoise-orm (~=0.24)",
    "uvicorn (~=0.35)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0"]
build-backend = "poetry.core.masonry.api"
